LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focus-02-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: element send_keys intercepted error

Harness Error (FAIL), message = element send_keys intercepted error

TIMEOUT The keydown event must be targeted at the body element Test timed out
NOTRUN The keypress event must be targeted at the body element
NOTRUN The keyup event must be targeted at the body element
